Improve `hijackMap` to work better with reloads
I think this code was relying on being run before the GeoGuessr page
loaded. With my changes, the renderer code is injected during/after the
load, so the page may already be ready. With these changes, we also
synchronously check for the existence of the google maps API script tag
as soon as the script is ran, so it can immediately resolve if it was
already there.