Fix settings window on Linux
On linux (and probably macOS) we need to use a `file://` URL for the
settings and update windows to load correctly.

Depending on the window manager, the settings window may be completely
closed, without us being able to prevent it. In that case the window
can't be shown again later, we need to recreate it.