Split country flags and non-country flags, add more unicode flags
This moves the flags for non-countries / subdivisions into a new file,
`builtinFlags.json`, which has a custom emoji field for display in chat
messages. The country flags will now show as "🇺🇸 (AK)" with the state
abbreviation instead of as 🏴󠁵󠁳󠁡󠁫󠁿.

In addition the `builtinFlags.json` file lets us add some other stuff that
has Unicode support but is not a country. This includes the white flag
and some pride flags, I think these would be nice to have. now people
can also finally do `!flag surrender` :P