rip out stats migration
Migrating all the existing statistics from the electron-store
implementation makes it very complicated to calculate the combined
stats. It's probably not worth the future maintenance burden. So for
now, let's not migrate old stats at all, but keep them in the
electron-store in case we have a reason to add a migration in the future.