Mark games as finished, fix #8
Now ongoing games have `state = 'started'`, while completed games have
`state = 'finished'`. This simplifies the queries that need to know game
winners, because they don't have to guess which games are finished based
on the number of rounds, which was wrong anyways.

Now when a game is finished, the GameHandler needs to call `finishGame`.