special-case some wrong streak counts
country-iso has very granular borders, but it doesn't seem to account
for holey polygons correctly, so countries enclosed by other countries
and enclaves get wrong results. This patch is a simple workaround around
the common cases where this happens. We just check some cases where we
get two results and pick one. In more complicated enclave/exclave
situations like in Baarle this is not possible to do. For a proper fix
we need to fix the library or use a different one that accounts for holes.