Don't allow using !me after guessing during a round
There is an exploit that can be used with the !me-command. After making a guess,
!me will include information about the guess
- Current streak will reflect whether user guessed the correct country
- Average score includes the score for the guess
- 5k count includes information about whether the user 5k'd.

If multiguess is turned on, then this information can be used to alter the guess!

This commit fixes this exploit by rejecting the !me-command after guessing during
an active round (active being determined whether the 'BROADCASTER' has guessed
or not).